express js added to package.json
